🐛⚡️Fixed & Improved cflearn.dist
It is well known that there are significant caveats to using CUDA models with multiprocessing, so this commit managed to resolve it in a tricky way based on `os.system`. Since `os.system` is usually not the best practice, please make sure in advance that you will not feel uncomfortable with the implementation of this commit before you utilize it.